Created toolkit uploader (#26)
* Created toolkit uploader

Created an uploader for the toolkit which uploads parsed JSON content to the Mongo database. Both replace and merge options are implemented.

* Updated uploader and database methods

Changed connectDB() to use non-deprecated methods, and fixed many issues in uploader (currently still broken)

* Adjust uploader code for IdWrapper removal

* Create generic upload function, clean up uploader.go

Created a generic function for uploading JSON files to Mongo, and added some comments for clarification. Currently the function requires that the file name is the same as the collection that the file contents are being uploaded to. Only courses, professors, and sections are supported.

* Delete data.7z

---------

Co-authored-by: jpahm <20374744+jpahm@users.noreply.github.com>